// src/models/regions/locations/context/types.ts
export interface LocationContext {
  idx: number
  tribal: boolean
  remote: boolean
  village: boolean
  urban: boolean
  city: boolean
  coastal: boolean
  stone: boolean
  lumber: boolean
  textiles: boolean
  mining: boolean
  herders: boolean
  grain: boolean
  vegetables: boolean
  incense: boolean
  arcane: boolean
  alchemy: boolean
  cosmetics: boolean
  mercenaries: boolean
  art: boolean
  text: boolean
  smiths: boolean
  gold: boolean
  woodwork: boolean
  jewels: boolean
  glass: boolean
  mechanics: boolean
  herbs: boolean
  ships: boolean
  clothing: boolean
  leather: boolean
  fur: boolean
  alcohol: boolean
  grapes: boolean
  ceramics: boolean
  bees: boolean
}